% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply.R
\name{multiply}
\alias{multiply}
\title{Multiplies two numbers}
\usage{
multiply(x, y)
}
\arguments{
\item{x, y}{Scalar values to multiply}
}
\description{
Performs multiplication of two integer or numeric scalars.
}
\examples{
multiply(2, 3)

}
\seealso{
\code{\link[=prod]{prod()}}
}
