% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tic.package-package.r
\docType{package}
\name{tic.package-package}
\alias{tic.package}
\alias{tic.package-package}
\title{tic.package: Example Package for 'tic'}
\description{
Building a package with pkgdown documentation.
}
\details{
For demonstration of the \pkg{tic} package.
}
\examples{
# Show timestamp
Sys.time()
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/krlmlr/tic.package}
  \item \url{http://krlmlr.github.io/tic.package}
  \item Report bugs at \url{https://github.com/krlmlr/tic.package/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

}
